Move dir_compare and file_compare to a new module
